refactor udfgenerator: remove consts.py
