--- mipengine/udfgen/consts.py
@@ -1,25 +1,0 @@
-MAIN_TABLE_PLACEHOLDER = "main_output_table_name"
-
-CREATE_OR_REPLACE_FUNCTION = "CREATE OR REPLACE FUNCTION"
-RETURNS = "RETURNS"
-LANGUAGE_PYTHON = "LANGUAGE PYTHON"
-BEGIN = "{"
-END = "}"
-
-SELECT = "SELECT"
-FROM = "FROM"
-WHERE = "WHERE"
-GROUP_BY = "GROUP BY"
-ORDER_BY = "ORDER BY"
-DROP_TABLE_IF_EXISTS = "DROP TABLE IF EXISTS"
-CREATE_TABLE = "CREATE TABLE"
-AND = "AND"
-
-SEP = ","
-LN = "\n"
-SEPLN = SEP + LN
-ANDLN = " " + AND + LN
-SPC4 = " " * 4
-SCOLON = ";"
-ROWID = "row_id"
-DEFERRED = "deferred"
